Added build file
One has to specify hostname they want to use but after they should be able to run it and it auto detect the cpu and motherboard values but I have to change name1 and name2 to the actual name or a maybe do a list of all possible names idk yet. I would like to have a build file though to do it for experience and to be able to pull it down from a repo and run it on any arch system with node and npm.